feat: Update package.json for merchant-app, store, and user-app
This commit updates the package.json files for the merchant-app, store, and user-app. It changes the "name" field in the merchant-app and user-app package.json files to reflect the correct app names. It also updates the exports in the store package.json file to correctly point to the useBalance hook file.

feat: Update route imports in auth and user APIs

This commit updates the route imports in the auth and user APIs of the user-app. It changes the import paths for the authOptions from the lib/auth file to the correct paths in the app's directory structure.

chore: Remove unused provider.tsx file

This commit removes the unused provider.tsx file from the user-app. The file was no longer needed and has been deleted.

feat: Add new components to ui package.json

This commit adds new components to the ui package.json file. It includes the center, select, and textinput components, with their respective file paths.

feat: Update page.tsx in user-app

This commit updates the page.tsx file in the user-app. It replaces the import statements for next-auth/react with the getServerSession function from next-auth. It also adds a redirect logic based on the user's session status.

feat: Update card.tsx in ui package

This commit updates the card.tsx file in the ui package. It removes the href prop from the component and replaces it with a div element. It also updates the title element to use an h1 tag instead of an h2 tag.

feat: Update layout.tsx in user-app

This commit updates the layout.tsx file in the user-app. It replaces the Providers component import with the Providers component from the components directory. It also adds an AppbarClient component to the layout.

feat: Update db package.json

This commit updates the db package.json file. It adds the bcrypt dependency and includes a prisma seed script.

feat: Update page.tsx in user-app

This commit updates the page.tsx file in the user-app. It replaces the import statements for next-auth/react with the getServerSession function from next-auth. It also adds a redirect logic based on the user's session status.

feat: Update card.tsx in ui package

This commit updates the card.tsx file in the ui package. It removes the href prop from the component and replaces it with a div element. It also updates the title element to use an h1 tag instead of an h2 tag.

feat: Update layout.tsx in user-app

This commit updates the layout.tsx file in the user-app. It replaces the Providers component import with the Providers component from the components directory. It also adds an AppbarClient component to the layout.